@@ -8,25 +8,25 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 
 dependencies:
   flutter:
     sdk: flutter
-  # pedantic: ^1.11.1
+  pedantic: ^1.11.1
 
 
-  cupertino_icons: ^1.0.2
-  path_provider: ^2.0.2
-  image_gallery_saver: ^1.6.9
+  cupertino_icons: ^1.0.3
+  path_provider: ^2.0.5
+  image_gallery_saver: ^1.7.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
 
-  # flutter_lints: ^1.0.0
+  flutter_lints: ^1.0.4
 
 
 flutter:
