@@ -2,7 +2,7 @@ import 'package:english_application/widgets/main_screen/drawing_page/drawing_pag
 import 'package:flutter/material.dart';
 
 class MyApp extends StatelessWidget {
-  const MyApp({Key key}) : super(key: key);
+  const MyApp({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
