@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 class Sketcher extends CustomPainter {
   final List<DrawnLine> lines;
 
-  Sketcher({this.lines});
+  Sketcher({required this.lines});
 
   @override
   void paint(Canvas canvas, Size size) {
@@ -14,6 +14,7 @@ class Sketcher extends CustomPainter {
       ..strokeWidth = 5.0;
 
     for (int i = 0; i < lines.length; ++i) {
+      // ignore: unnecessary_null_comparison
       if (lines[i] == null) continue;
       for (int j = 0; j < lines[i].path.length - 1; ++j) {
         if (lines[i].path[j] != null && lines[i].path[j + 1] != null) {
